% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addPoints}
\alias{addPoints}
\title{Add time series of points to an existing xts plot}
\usage{
addPoints(x, main = "", on = NA, col = NULL, pch = 0, ...)
}
\arguments{
\item{x}{an xts object to plot.}

\item{main}{main title for a new panel if drawn.}

\item{on}{panel number to draw on. A new panel will be drawn if \code{on=NA}.}

\item{col}{color palette to use, set by default to rational choices.}

\item{pch}{the type of plot to be drawn, same as in \code{\link{plot}}.}

\item{\dots}{any other passthrough parameters. Not currently used.}
}
\description{
Add time series of points to an existing xts plot
}
\author{
Ross Bennett
}

