% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date.R, R/POSIX.R, R/data.frame.R, R/irts.R,
%   R/matrix.R, R/reclass.R, R/timeDate.R, R/timeSeries.R, R/ts.R, R/xts.R,
%   R/yearmon.R, R/zoo.R
\name{as.xts.Date}
\alias{as.xts.Date}
\alias{as.xts.POSIXt}
\alias{as.xts.data.frame}
\alias{as.xts.irts}
\alias{as.xts.matrix}
\alias{try.xts}
\alias{use.xts}
\alias{Reclass}
\alias{use.reclass}
\alias{as.xts.timeDate}
\alias{as.xts.timeSeries}
\alias{as.xts.ts}
\alias{as.xts}
\alias{xtsible}
\alias{as.xts.yearmon}
\alias{as.xts.yearqtr}
\alias{as.xts.zoo}
\title{Convert Object To And From Class xts}
\usage{
\method{as.xts}{Date}(x, ...)

\method{as.xts}{POSIXt}(x, ...)

\method{as.xts}{data.frame}(
  x,
  order.by,
  dateFormat = "POSIXct",
  frequency = NULL,
  ...,
  .RECLASS = FALSE
)

\method{as.xts}{irts}(x, order.by, frequency = NULL, ..., .RECLASS = FALSE)

\method{as.xts}{matrix}(
  x,
  order.by,
  dateFormat = "POSIXct",
  frequency = NULL,
  ...,
  .RECLASS = FALSE
)

try.xts(x, ..., error = TRUE)

Reclass(x)

\method{as.xts}{timeDate}(x, ...)

\method{as.xts}{timeSeries}(
  x,
  dateFormat = "POSIXct",
  FinCenter,
  recordIDs,
  title,
  documentation,
  ...,
  .RECLASS = FALSE
)

\method{as.xts}{ts}(x, dateFormat, ..., .RECLASS = FALSE)

as.xts(x, ...)

xtsible(x)

\method{as.xts}{yearmon}(x, ...)

\method{as.xts}{yearqtr}(x, ...)

\method{as.xts}{zoo}(x, order.by = index(x), frequency = NULL, ..., .RECLASS = FALSE)
}
\arguments{
\item{x}{data object to convert. See details for supported types}

\item{\dots}{additional parameters or attributes}

\item{order.by}{see \link[zoo:zoo]{zoo} help}

\item{dateFormat}{what format should the dates be converted to}

\item{frequency}{see \link[zoo:zoo]{zoo} help}

\item{.RECLASS}{should conversion be reversible?}

\item{error}{error handling option. See Details.}

\item{FinCenter}{see timeSeries help}

\item{recordIDs}{see timeSeries help}

\item{title}{see timeSeries help}

\item{documentation}{see timeSeries help}
}
\value{
An S3 object of class \code{xts}.
}
\description{
Conversion S3 methods to coerce data objects of arbitrary classes to class
\code{xts} and back, without losing any attributes of the original format.
}
\details{
A simple and reliable way to convert many different objects into a uniform
format for use within \R.

It is possible with a call to \code{as.xts} to convert objects of class
\code{timeSeries}, \code{ts}, \code{matrix}, \code{data.frame}, and
\code{zoo}.

Additional name=value pairs may be passed to the function to be added to the
new object. A special print.xts method will assure that the attributes are
hidden from view, but will be available via \R's standard \code{attr}
function.

If \code{.RECLASS=TRUE}, the returned object will preserve all relevant
attribute/slot data within itself, allowing for temporary conversion to use
zoo and xts compatible methods. A call to \code{reclass} returns the object
to its original class, with all original attributes intact - unless
otherwise changed.  This is the default behavior when \code{try.xts} is used
for conversion, and should not be altered by the user; i.e. don't touch it
unless you are aware of the consequences.

It should be obvious, but any attributes added via the \dots argument will
not be carried back to the original data object, as there would be no
available storage slot/attribute.
}
\examples{

  \dontrun{
  # timeSeries
  library(timeSeries)
  x <- timeSeries(1:10, 1:10)

  str( as.xts(x) )
  str( reclass(as.xts(x)) )
  str( try.xts(x) )
  str( reclass(try.xts(x)) )
  }

}
\seealso{
\code{\link[=xts]{xts()}}, \code{\link[zoo:zoo]{zoo::zoo()}}
}
\author{
Jeffrey A. Ryan
}
\keyword{utilities}
