% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts.R
\name{CLASS}
\alias{CLASS}
\alias{CLASS<-}
\title{Extract and Set .CLASS Attribute}
\usage{
CLASS(x)

CLASS(x) <- value
}
\arguments{
\item{x}{an xts object}

\item{value}{the new .CLASS value to assign}
}
\value{
Called for its side-effect of changing the .CLASS attribute
}
\description{
Simple extraction and replacement function to access \code{xts} .CLASS
attribute.  The .CLASS attribute is used by \code{reclass} to transform an
\code{xts} object back to its original class.
}
\details{
It is not recommended that CLASS be called in daily use.  While it may be
possible to coerce objects to other classes than originally derived from,
there is little, if any, chance that the \code{reclass} function will
perform as expected.

It is best to use the traditional \code{as} methods.
}
\seealso{
\code{\link{as.xts}},\code{\link{reclass}}
}
\author{
Jeffrey A. Ryan
}
\keyword{utilities}
