% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{firstof}
\alias{firstof}
\alias{lastof}
\title{Create a POSIXct Object}
\usage{
firstof(year = 1970, month = 1, day = 1, hour = 0, min = 0, sec = 0, tz = "")

lastof(
  year = 1970,
  month = 12,
  day = 31,
  hour = 23,
  min = 59,
  sec = 59,
  subsec = 0.99999,
  tz = ""
)
}
\arguments{
\item{year, month, day}{Numeric values to specify a day.}

\item{hour, min, sec}{Numeric vaues to specify time within a day.}

\item{tz}{Timezone used for conversion.}

\item{subsec}{Number of sub-seconds.}
}
\value{
An POSIXct object.
}
\description{
Easily create of time stamps corresponding to the first or last observation
in a specified time period.
}
\details{
This is a wrapper to \code{\link[=ISOdatetime]{ISOdatetime()}} with defaults corresponding to the
first or last possible time in a given period.
}
\examples{

firstof(2000)
firstof(2005,01,01)

lastof(2007)
lastof(2007,10)

}
\seealso{
\code{\link[=ISOdatetime]{ISOdatetime()}}
}
\author{
Jeffrey A. Ryan
}
\keyword{utilities}
