% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{period.min}
\alias{period.min}
\title{Calculate Min By Period}
\usage{
period.min(x, INDEX)
}
\arguments{
\item{x}{a univariate data object}

\item{INDEX}{a numeric vector of endpoints to calculate maximum on}
}
\value{
An xts or zoo object of minimums, indexed by the period endpoints.
}
\description{
Calculate a minimum for each period of INDEX. Essentially a rolling
application of minimum over a series of non-overlapping sections.
}
\details{
Used to calculate a minimum per period given an arbitrary index of sections
to be calculated over. This is an optimized function for minimum. There are
additional optimized versions for max, sum, and prod.

For xts-coercible objects, an appropriate INDEX can be derived from a call
to \code{endpoints}.
}
\examples{

period.min(c(1,1,4,2,2,6,7,8,-1,20),c(0,3,5,8,10))

data(sample_matrix)
period.min(sample_matrix[,1],endpoints(sample_matrix))
period.min(as.xts(sample_matrix)[,1],endpoints(sample_matrix))

}
\seealso{
\code{\link{endpoints}}, \code{\link{period.sum}},
\code{\link{period.max}}, \code{\link{period.prod}}
}
\author{
Jeffrey A. Ryan
}
\keyword{utilities}
