% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{period.sum}
\alias{period.sum}
\title{Calculate Sum By Period}
\usage{
period.sum(x, INDEX)
}
\arguments{
\item{x}{a univariate data object}

\item{INDEX}{a numeric vector of endpoints to calculate sum on}
}
\value{
An \code{xts} or \code{zoo} object of sums, indexed by the period
endpoints.
}
\description{
Calculate a sum for each period of INDEX. Essentially a rolling application
of sum over a series of non-overlapping sections.
}
\details{
Used to calculate a sum per period given an arbitrary index of sections to
be calculated over. This is an optimized function for sum.  There are
additionally optimized versions for min, max, and prod.

For xts-coercible objects, an appropriate INDEX can be derived from a call
to \code{endpoints}.
}
\examples{

period.sum(c(1,1,4,2,2,6,7,8,-1,20),c(0,3,5,8,10))

data(sample_matrix)
period.sum(sample_matrix[,1],endpoints(sample_matrix))
period.sum(as.xts(sample_matrix)[,1],endpoints(sample_matrix))

}
\seealso{
\code{\link[=endpoints]{endpoints()}}, \code{\link[=period.max]{period.max()}},
\code{\link[=period.min]{period.min()}}, \code{\link[=period.prod]{period.prod()}}
}
\author{
Jeffrey A. Ryan
}
\keyword{utilities}
