% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts-package.R
\docType{data}
\name{sample_matrix}
\alias{sample_matrix}
\title{Sample Data Matrix For xts Example and Unit Testing}
\format{
\preformatted{The format is:
  num [1:180, 1:4] 50.0 50.2 50.4 50.4 50.2 ...
  - attr(*, "dimnames")=List of 2
    ..$ : chr [1:180] "2007-01-02" "2007-01-03" "2007-01-04" "2007-01-05" ...
    ..$ : chr [1:4] "Open" "High" "Low" "Close" }
}
\usage{
data(sample_matrix)
}
\description{
Simulated 180 observations on 4 variables.
}
\examples{

data(sample_matrix)

}
\keyword{datasets}
