% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split.xts}
\alias{split.xts}
\title{Divide into Groups by Time}
\usage{
\method{split}{xts}(x, f = "months", drop = FALSE, k = 1, ...)
}
\arguments{
\item{x}{an xts object}

\item{f}{a 'character' vector describing the period to split by}

\item{drop}{ignored by split.xts}

\item{k}{number of periods to aggregate into each split. See Details.}

\item{\dots}{further args to non-xts method}
}
\value{
A list of xts objects.
}
\description{
Creates a list of xts objects split along time periods.
}
\details{
A quick way to break up a large xts object by standard time periods; e.g.
'months', 'quarters', etc.

\code{endpoints} is used to find the start and end of each period (or
k-periods).  See that function for valid arguments.

If \code{f} is not a character vector, the NextMethod is called, which would
in turn dispatch to the split.zoo method.
}
\note{
\code{aggregate.zoo} would be more flexible, though not as fast for
xts objects.
}
\examples{

data(sample_matrix)
x <- as.xts(sample_matrix)


split(x)
split(x, f="weeks")
split(x, f="weeks", k=4)

}
\seealso{
\code{\link{endpoints}}, \code{\link[zoo]{split.zoo}},
\code{\link[zoo]{aggregate.zoo}}
}
\author{
Jeffrey A. Ryan
}
\keyword{utilities}
