% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{is.timeBased}
\alias{is.timeBased}
\alias{timeBased}
\title{Check if Class is Time-Based}
\usage{
is.timeBased(x)

timeBased(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
A logical scalar.
}
\description{
Used to verify that the object is one of the known time-based classes in R.
Current time-based objects supported are \code{Date}, \code{POSIXct}, \code{chron},
\code{yearmon}, \code{yearqtr}, and \code{timeDate}.
}
\examples{

timeBased(Sys.time())
timeBased(Sys.Date())

timeBased(200701)

}
\author{
Jeffrey A. Ryan
}
\keyword{utilities}
