% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{addLegend}
\alias{addLegend}
\title{Add Legend}
\usage{
addLegend(
  legend.loc = "topright",
  legend.names = NULL,
  col = NULL,
  ncol = 1,
  on = 0,
  ...
)
}
\arguments{
\item{legend.loc}{One of nine locations: bottomright, bottom, bottomleft,
left, topleft, top, topright, right, or center.}

\item{legend.names}{Character vector of names for the legend. When \code{NULL},
the column names of the current plot object are used.}

\item{col}{Fill colors for the legend. When \code{NULL}, the colorset of the
current plot object data is used.}

\item{ncol}{Number of columns for the legend.}

\item{on}{Panel number to draw on. A new panel will be drawn if \code{on = NA}.
The default, \code{on = 0}, will add to the active panel. The active panel is
defined as the panel on which the most recent action was performed. Note
that only the first element of \code{on} is checked for the default behavior to
add to the last active panel.}

\item{\dots}{Any other passthrough parameters to \code{\link[=legend]{legend()}}.}
}
\description{
Add a legend to an existing panel.
}
\author{
Ross Bennett
}
