% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{addSeries}
\alias{addSeries}
\title{Add a time series to an existing xts plot}
\usage{
addSeries(
  x,
  main = "",
  on = NA,
  type = "l",
  col = NULL,
  lty = 1,
  lwd = 1,
  pch = 1,
  ...
)
}
\arguments{
\item{x}{An xts object to add to the plot.}

\item{main}{Main title for a new panel if drawn.}

\item{on}{Panel number to draw on. A new panel will be drawn if \code{on = NA}.}

\item{type}{The type of plot to be drawn, same as in \code{\link[=plot]{plot()}}.}

\item{col}{Color palette to use, set by default to rational choices.}

\item{lty}{Set the line type, same as in \code{\link[=par]{par()}}.}

\item{lwd}{Set the line width, same as in \code{\link[=par]{par()}}.}

\item{pch}{The type of plot to be drawn, same as in \code{\link[=par]{par()}}.}

\item{\dots}{Any other passthrough graphical parameters.}
}
\description{
Add a time series to an existing xts plot
}
\author{
Ross Bennett
}
