\name{sample_matrix}
\alias{sample_matrix}
\docType{data}
\title{ Sample Data Matrix For xts Example and Unit Testing }
\description{
Simulated 180 observations on 4 variables.
}
\usage{data(sample_matrix)}
\format{
\preformatted{
  The format is:
 num [1:180, 1:4] 50.0 50.2 50.4 50.4 50.2 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:180] "2007-01-02" "2007-01-03" "2007-01-04" "2007-01-05" ...
  ..$ : chr [1:4] "Open" "High" "Low" "Close"
}
}
\examples{
data(.sample.matrix)
}
\keyword{datasets}
