% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts-package.R
\docType{package}
\name{xts-package}
\alias{xts-package}
\title{xts: extensible time-series}
\description{
Extensible time series class and methods, extending and behaving like zoo.
}
\details{
Easily convert one of \R's many time-series (and non-time-series) classes to a
true time-based object which inherits all of zoo's methods, while allowing
for new time-based tools where appropriate.

Additionally, one may use \pkg{xts} to create new objects which can contain
arbitrary attributes named during creation as name=value pairs.
}
\seealso{
\code{\link[=xts]{xts()}}, \code{\link[=as.xts]{as.xts()}}, \code{\link[=reclass]{reclass()}}, \code{\link[zoo:zoo]{zoo()}}
}
\author{
Jeffrey A. Ryan and Joshua M. Ulrich

Maintainer: Joshua M. Ulrich \href{mailto:josh.m.ulrich@gmail.com}{josh.m.ulrich@gmail.com}
}
\keyword{package}
